+++ src/Bizagi.Microservice.Api.WorldCupPool.Abstractions/Services/ILeaderboardService.cs
@@ -20,5 +20,5 @@ public interface ILeaderboardService
     /// </summary>
     /// <param name="userId">Identificador del usuario cuyo historial se consulta.</param>
     /// <returns>Lista de entradas del historial del usuario.</returns>
-    Task<IEnumerable<UserPredictionHistoryDto>> GetUserHistoryAsync(decimal userId);
+    Task<IEnumerable<UserPredictionHistoryDto>> GetUserHistoryAsync(int userId);
 }
