+++ src/Bizagi.Microservice.Api.WorldCupPool.Data.Transfer.Object/Admin/MatchResultDto.cs
@@ -8,12 +8,12 @@ public class MatchResultDto
     /// <summary>
     /// Identificador único del resultado registrado.
     /// </summary>
-    public decimal Id { get; set; }
+    public int Id { get; set; }
 
     /// <summary>
     /// Identificador del partido al que corresponde el resultado.
     /// </summary>
-    public decimal IdMatch { get; set; }
+    public int IdMatch { get; set; }
 
     /// <summary>
     /// Goles reales del equipo local.
@@ -28,7 +28,7 @@ public class MatchResultDto
     /// <summary>
     /// Identificador del administrador que registró el resultado.
     /// </summary>
-    public decimal RegisteredByUserId { get; set; }
+    public int RegisteredByUserId { get; set; }
 
     /// <summary>
     /// Fecha y hora en que se registró el resultado.
